+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -141,7 +141,7 @@ class VideoModule : Module() {
       Constructor { source: VideoSource ->
         val mediaItem = source.toMediaItem()
         VideoManager.registerVideoSourceToMediaItem(mediaItem, source)
-        VideoPlayer(activity.applicationContext, appContext, mediaItem)
+        VideoPlayer(activity.applicationContext, appContext, mediaItem, source)
       }
 
       Property("playing")
@@ -206,6 +206,16 @@ class VideoModule : Module() {
           }
         }
 
+      Property("showNowPlayingNotification")
+        .get { ref: VideoPlayer ->
+          ref.showNowPlayingNotification
+        }
+        .set { ref: VideoPlayer, showNotification: Boolean ->
+          appContext.mainQueue.launch {
+            ref.showNowPlayingNotification = showNotification
+          }
+        }
+
       Property("status")
         .get { ref: VideoPlayer ->
           ref.status
@@ -255,6 +265,7 @@ class VideoModule : Module() {
         VideoManager.registerVideoSourceToMediaItem(mediaItem, videoSource)
 
         appContext.mainQueue.launch {
+          ref.videoSource = videoSource
           ref.player.setMediaItem(mediaItem)
         }
       }